MYSQL Steps

+++ Vision-Auto-Tests/Vision-Tests/src/test/resources/Features/IPv6/IPv6 export Device Configuration.feature
@@ -11,11 +11,11 @@ Feature: IPv6 Export device configuration
 #    Then CLI Run remote linux Command "mysql -prad123 vision_ng -e "delete from device_file where dev_type='Alteon';"" on "ROOT_SERVER_CLI"
     Then MYSQL DELETE FROM "device_file" Table in "VISION_NG" Schema WHERE "dev_type='Alteon'"
 
-    Then CLI Run remote linux Command "mysql -prad123 vision_ng -e "delete from site_tree_elem_abs where name='Alteon_200a::172:17:164:19';"" on "ROOT_SERVER_CLI"
-    Then MYSQL DELETE FROM "" Table in "VISION_NG" Schema WHERE ""
+#    Then CLI Run remote linux Command "mysql -prad123 vision_ng -e "delete from site_tree_elem_abs where name='Alteon_200a::172:17:164:19';"" on "ROOT_SERVER_CLI"
+    Then MYSQL DELETE FROM "site_tree_elem_abs" Table in "VISION_NG" Schema WHERE "name='Alteon_200a::172:17:164:19'"
 
     Then CLI Run remote linux Command "mysql -prad123 vision_ng -e "delete from device_setup where fk_dev_access_device_acces=(select row_id from device_access where mgt_ip="200a:0:0:0:172:17:164:19");"" on "ROOT_SERVER_CLI"
-    Then MYSQL DELETE FROM "" Table in "VISION_NG" Schema WHERE ""
+    Then MYSQL DELETE FROM "device_setup" Table in "VISION_NG" Schema WHERE "fk_dev_access_device_acces=(select row_id from device_access where mgt_ip='200a:0:0:0:172:17:164:19')"
 
     Then CLI Run remote linux Command "mysql -prad123 vision_ng -e "delete from device_access where mgt_ip="200a:0:0:0:172:17:164:19";"" on "ROOT_SERVER_CLI"
     Then MYSQL DELETE FROM "" Table in "VISION_NG" Schema WHERE ""
